Update Roles, Permission, Master User, Master Warehouse

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -10,17 +10,23 @@ class PermissionController extends Controller
 {
     public function index()
     {
-        $permissions = Permission::get();
+        abort_unless(\Gate::allows('Permission Access'), 403);
+
+        $permissions = Permission::orderBy('name', 'ASC')->get();
         return view('admin.permission.index', compact('permissions'));
     }
 
     public function create()
     {
+        abort_unless(\Gate::allows('Permission Create'), 403);
+
         return view('admin.permission.create');
     }
 
     public function store(StorePermissionRequest $request)
     {
+        abort_unless(\Gate::allows('Permission Create'), 403);
+
         Permission::create($request->all());
 
         return redirect()->route('admin.permissions.index')->with('success', 'Permission has been added');
@@ -29,11 +35,15 @@ class PermissionController extends Controller
 
     public function edit(Permission $permission)
     {
+        abort_unless(\Gate::allows('Permission Edit'), 403);
+
         return view('admin.permission.edit', compact('permission'));
     }
 
     public function update(UpdatePermissionRequest $request, Permission $permission)
     {
+        abort_unless(\Gate::allows('Permission Edit'), 403);
+
         $permission->update($request->all());
 
         return redirect()->route('admin.permissions.index')->with('success', 'Permission has been updated');
@@ -41,6 +51,8 @@ class PermissionController extends Controller
 
     public function destroy(Permission $permission)
     {
+        abort_unless(\Gate::allows('Permission Delete'), 403);
+
         $permission->delete();
 
         return redirect()->route('admin.permissions.index')->with('success', 'Permission has been deleted');
